Move S3Uploader out into its own file
